Moved util
